--- toggles-app/src/main/java/com/izettle/wrench/dialogs/DialogWidth.kt
@@ -1,20 +1,0 @@
-package com.izettle.wrench.dialogs
-
-import android.content.res.Resources
-import android.graphics.Rect
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
-
-/**
- * https://stackoverflow.com/a/41495370
- * Call this method (in onActivityCreated or later) to set
- * the width of the dialog to a percentage of the current
- * screen width.
- */
-fun DialogFragment.setWidthPercent(percentage: Int) {
-    val percent = percentage.toFloat() / 100
-    val dm = Resources.getSystem().displayMetrics
-    val rect = dm.run { Rect(0, 0, widthPixels, heightPixels) }
-    val percentWidth = rect.width() * percent
-    dialog?.window?.setLayout(percentWidth.toInt(), ViewGroup.LayoutParams.WRAP_CONTENT)
-}
