+++ toggles-app/src/main/java/se/eelde/toggles/dialogs/booleanvalue/BooleanValueFragment.kt
@@ -1,4 +1,4 @@
-package com.izettle.wrench.dialogs.stringvalue
+package se.eelde.toggles.dialogs.booleanvalue
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,15 +8,15 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.findNavController
+import com.izettle.wrench.dialogs.booleanvalue.FragmentBooleanValueViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import se.eelde.toggles.R
 import se.eelde.toggles.TogglesTheme
-import se.eelde.toggles.dialogs.stringvalue.StringValueView
 
 @AndroidEntryPoint
-class StringValueFragment : Fragment() {
+class BooleanValueFragment : Fragment() {
 
-    private val viewModel by viewModels<FragmentStringValueViewModel>()
+    private val viewModel by viewModels<FragmentBooleanValueViewModel>()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -25,7 +25,7 @@ class StringValueFragment : Fragment() {
     ): View = inflater.inflate(R.layout.fragment_compose, container, false).apply {
         findViewById<ComposeView>(R.id.compose).setContent {
             TogglesTheme {
-                StringValueView(findNavController(), viewModel)
+                BooleanValueView(findNavController(), viewModel)
             }
         }
     }
