+++ toggles-app/src/main/java/se/eelde/toggles/dialogs/enumvalue/EnumValueFragment.kt
@@ -1,4 +1,4 @@
-package com.izettle.wrench.dialogs.enumvalue
+package se.eelde.toggles.dialogs.enumvalue
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,18 +8,15 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.findNavController
-import androidx.navigation.fragment.navArgs
+import com.izettle.wrench.dialogs.enumvalue.FragmentEnumValueViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import se.eelde.toggles.R
 import se.eelde.toggles.TogglesTheme
-import se.eelde.toggles.dialogs.enumvalue.EnumValueView
 
 @AndroidEntryPoint
 class EnumValueFragment : Fragment() {
 
     private val viewModel by viewModels<FragmentEnumValueViewModel>()
-
-    private val args: EnumValueFragmentArgs by navArgs()
 
     override fun onCreateView(
         inflater: LayoutInflater,
