+++ toggles-app/src/main/java/se/eelde/toggles/dialogs/stringvalue/StringValueFragment.kt
@@ -1,4 +1,4 @@
-package com.izettle.wrench.dialogs.booleanvalue
+package se.eelde.toggles.dialogs.stringvalue
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,15 +8,15 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.findNavController
+import com.izettle.wrench.dialogs.stringvalue.FragmentStringValueViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import se.eelde.toggles.R
 import se.eelde.toggles.TogglesTheme
-import se.eelde.toggles.dialogs.booleanvalue.BooleanValueView
 
 @AndroidEntryPoint
-class BooleanValueFragment : Fragment() {
+class StringValueFragment : Fragment() {
 
-    private val viewModel by viewModels<FragmentBooleanValueViewModel>()
+    private val viewModel by viewModels<FragmentStringValueViewModel>()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -25,7 +25,7 @@ class BooleanValueFragment : Fragment() {
     ): View = inflater.inflate(R.layout.fragment_compose, container, false).apply {
         findViewById<ComposeView>(R.id.compose).setContent {
             TogglesTheme {
-                BooleanValueView(findNavController(), viewModel)
+                StringValueView(findNavController(), viewModel)
             }
         }
     }
