+++ toggles-app/src/main/res/navigation/application_configurations_graph.xml
@@ -46,7 +46,7 @@
     </fragment>
     <fragment
         android:id="@+id/enumValueFragment"
-        android:name="com.izettle.wrench.dialogs.enumvalue.EnumValueFragment"
+        android:name="se.eelde.toggles.dialogs.enumvalue.EnumValueFragment"
         android:label="EnumValueFragment">
         <argument
             android:name="configurationId"
@@ -57,7 +57,7 @@
     </fragment>
     <fragment
         android:id="@+id/stringValueFragment"
-        android:name="com.izettle.wrench.dialogs.stringvalue.StringValueFragment"
+        android:name="se.eelde.toggles.dialogs.stringvalue.StringValueFragment"
         android:label="StringValueFragment">
         <argument
             android:name="configurationId"
@@ -68,7 +68,7 @@
     </fragment>
     <fragment
         android:id="@+id/integerValueFragment"
-        android:name="com.izettle.wrench.dialogs.integervalue.IntegerValueFragment"
+        android:name="se.eelde.toggles.dialogs.integervalue.IntegerValueFragment"
         android:label="IntegerValueFragment">
         <argument
             android:name="configurationId"
@@ -79,7 +79,7 @@
     </fragment>
     <fragment
         android:id="@+id/booleanValueFragment"
-        android:name="com.izettle.wrench.dialogs.booleanvalue.BooleanValueFragment"
+        android:name="se.eelde.toggles.dialogs.booleanvalue.BooleanValueFragment"
         android:label="BooleanValueFragment">
         <argument
             android:name="configurationId"
