+++ toggles-app/src/main/res/navigation/navigation_graph.xml
@@ -48,7 +48,7 @@
 
     <fragment
         android:id="@+id/action_help"
-        android:name="com.izettle.wrench.help.HelpFragment"
+        android:name="se.eelde.toggles.help.HelpFragment"
         android:label="@string/help"
         tools:layout="@layout/fragment_compose" />
